LSPS1: Simplify and align client logic
.. we tremendously simplify the client logic, also dropping any state
tracking for now (possibly to be re-added in the future).

We also get the LSPS1 API closer to LSPS2 and make sure the API/docs is
more consistent.

To this end, we also rename a bunch of types
(`OrderParams`->`OrderParameters`, `OptionsSupported`->`LSPS1Options`)
for improved clarity and, since we're not using them, drop the
`AChannelManager` and `Filter` arguments of `LSPS1ClientHandler`.